+++ cf/test/test_Data.py
@@ -100,7 +100,7 @@ class DataTest(unittest.TestCase):
 #        ]
         
 #        self.test_only = ['test_Data_mean_mean_absolute_value']
-#        self.test_only = ['test_Data_AUXILIARY_MASK']
+        self.test_only = ['test_Data_AUXILIARY_MASK']
 #        self.test_only = ['test_Data_mean_of_upper_decile']
 #        self.test_only = ['test_Data__collapse_SHAPE']
 #        self.test_only = ['test_Data__collapse_UNWEIGHTED_MASKED']